Revert "SFA-1478: Base_rate vardef fix."
This reverts commit fe3c67f645b286632f2c56e7539cf5b43067ffc4.

+++ sugarcrm/modules/ForecastManagerWorksheets/vardefs.php
@@ -70,7 +70,7 @@ $dictionary['ForecastManagerWorksheet'] = array(
         array(
             'name' => 'base_rate',
             'vname' => 'LBL_BASE_RATE',
-            'type' => 'decimal',
+            'type' => 'double',
         ),
         'timeperiod_id' =>
         array(
